add webpack builder. Update service files. Re-write modules in es6 syntax
