+++ public/index.html
@@ -41,9 +41,7 @@
           </section>
       </div>
     </main>
-    <script src="../src/tripDays.js"></script>
-    <script src="../src/tripEvents.js"></script>
-    <script src="../src/tripMenu.js"></script>
-    <script src="../src/main.js"></script>
+    <script src="./js/bundle.js"></script>
+
   </body>
 </html>
