--- src/tripMenu.js
@@ -1,69 +1,0 @@
-'use strict';
-
-(function () {
-  function createInfoTemplate() {
-    return (
-      `<section class="trip-main__trip-info  trip-info">
-        <div class="trip-info__main">
-          <h1 class="trip-info__title">Amsterdam &mdash; Chamonix &mdash; Geneva</h1>
-
-          <p class="trip-info__dates">Mar 18&nbsp;&mdash;&nbsp;20</p>
-        </div>
-
-        <p class="trip-info__cost">
-          Total: &euro;&nbsp;<span class="trip-info__cost-value">1230</span>
-        </p>
-      </section>`
-    );
-  }
-
-  function createFiltersTemplate() {
-    return (
-      `<form class="trip-filters" action="#" method="get">
-      <div class="trip-filters__filter">
-        <input id="filter-everything" class="trip-filters__filter-input  visually-hidden" type="radio" name="trip-filter" value="everything" checked>
-        <label class="trip-filters__filter-label" for="filter-everything">Everything</label>
-      </div>
-
-      <div class="trip-filters__filter">
-        <input id="filter-future" class="trip-filters__filter-input  visually-hidden" type="radio" name="trip-filter" value="future">
-        <label class="trip-filters__filter-label" for="filter-future">Future</label>
-      </div>
-
-      <div class="trip-filters__filter">
-        <input id="filter-past" class="trip-filters__filter-input  visually-hidden" type="radio" name="trip-filter" value="past">
-        <label class="trip-filters__filter-label" for="filter-past">Past</label>
-      </div>
-
-      <button class="visually-hidden" type="submit">Accept filter</button>
-    </form>`
-    );
-  }
-
-  function createNavTabsTemplate() {
-    return (
-      `<nav class="trip-controls__trip-tabs  trip-tabs">
-        <a class="trip-tabs__btn  trip-tabs__btn--active" href="#">Table</a>
-        <a class="trip-tabs__btn" href="#">Stats</a>
-      </nav>`
-    );
-  }
-
-  function createControlsTemplate() {
-    return (
-      `${createInfoTemplate()}
-        <div class="trip-main__trip-controls  trip-controls">
-          <h2 class="visually-hidden">Switch trip view</h2>
-            ${createNavTabsTemplate()}
-          <h2 class="visually-hidden">Filter events</h2>
-            ${createFiltersTemplate()}
-        </div>`
-    );
-  }
-
-
-  window.tripMenu = {
-    createControlsTemplate
-  };
-
-})();
